Suppress warning when using release version.
When running protein_mpnn_run.py on a non-Git copy (e.g. the released versions), a rather concerning error message gets printed:
    
    fatal: Not a git repository: '/path/to/ProteinMPNN/.git'

Despite the message, this is not a fatal error for ProteinMPNN. 

Since the script handles the issue, we can simply suppress the printing of the error by discarding stderr.

There probably should be a better mechanism to insert the "official" release version here, but failing that suppressing the error message is probably the more user friendly way of doing things.